myproject/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9        # драйвер PostgreSQL
django-redis>=5.4           # кэш через Redis
gunicorn>=22.0              # опционально для прод-режима